+++ frontend/src/features/data_entry/components/voters_and_votes/VotersAndVotesForm.test.tsx
@@ -26,7 +26,7 @@ import { DataEntryState } from "../../types/types";
 import { getClientState } from "../../utils/dataEntryUtils";
 import { DataEntryProvider } from "../DataEntryProvider";
 import { VotersAndVotesForm } from "./VotersAndVotesForm";
-import { VotersAndVotesValues } from "./votersAndVotesValues";
+
 
 const initialValues: PollingStationResults = {
   recounted: undefined,
@@ -83,14 +83,9 @@ const defaultDataEntryState: DataEntryState = {
   cache: null,
 };
 
-function renderForm(initialDataEntryState?: DataEntryState) {
+function renderForm() {
   return render(
-    <DataEntryProvider
-      election={electionMockData}
-      pollingStationId={1}
-      entryNumber={1}
-      initialDataEntryState={initialDataEntryState}
-    >
+    <DataEntryProvider election={electionMockData} pollingStationId={1} entryNumber={1}>
       <VotersAndVotesForm />
     </DataEntryProvider>,
   );
@@ -374,54 +369,6 @@ describe("Test VotersAndVotesForm", () => {
       const request_body = body as POLLING_STATION_DATA_ENTRY_SAVE_REQUEST_BODY;
       expect(request_body.data).toEqual(expectedRequest.data);
     });
-
-    test("VotersAndVotesForm with cache", async () => {
-      const cacheData: VotersAndVotesValues = {
-        voters_counts: {
-          poll_card_count: 100,
-          proxy_certificate_count: 200,
-          voter_card_count: 300,
-          total_admitted_voters_count: 600,
-        },
-        votes_counts: {
-          votes_candidates_count: 400,
-          blank_votes_count: 500,
-          invalid_votes_count: 600,
-          total_votes_cast_count: 150,
-        },
-      };
-      renderForm({
-        ...defaultDataEntryState,
-        cache: {
-          key: "voters_votes_counts",
-          data: cacheData,
-        },
-      });
-      const pollCards = await screen.findByRole("textbox", { name: "A Stempassen" });
-      expect(pollCards).toHaveValue(cacheData.voters_counts.poll_card_count.toString());
-
-      const proxyCertificates = screen.getByRole("textbox", { name: "B Volmachtbewijzen" });
-      expect(proxyCertificates).toHaveValue(cacheData.voters_counts.proxy_certificate_count.toString());
-
-      const voterCards = screen.getByRole("textbox", { name: "C Kiezerspassen" });
-      expect(voterCards).toHaveValue(cacheData.voters_counts.voter_card_count.toString());
-
-      const totalAdmittedVoters = screen.getByRole("textbox", { name: "D Totaal toegelaten kiezers" });
-      expect(totalAdmittedVoters).toHaveValue(cacheData.voters_counts.total_admitted_voters_count.toString());
-
-      const votesOnCandidates = screen.getByRole("textbox", { name: "E Stemmen op kandidaten" });
-      expect(votesOnCandidates).toHaveValue(cacheData.votes_counts.votes_candidates_count.toString());
-
-      const blankVotes = screen.getByRole("textbox", { name: "F Blanco stemmen" });
-      expect(blankVotes).toHaveValue(cacheData.votes_counts.blank_votes_count.toString());
-
-      const invalidVotes = screen.getByRole("textbox", { name: "G Ongeldige stemmen" });
-      expect(invalidVotes).toHaveValue(cacheData.votes_counts.invalid_votes_count.toString());
-
-      const totalVotesCast = screen.getByRole("textbox", { name: "H Totaal uitgebrachte stemmen" });
-      expect(totalVotesCast).toHaveValue(cacheData.votes_counts.total_votes_cast_count.toString());
-    });
-  });
 
   describe("VotersAndVotesForm errors", () => {
     test("F.201 IncorrectTotal Voters counts", async () => {
