+++ frontend/src/features/data_entry/hooks/useDataEntry.ts
@@ -7,7 +7,7 @@ import {
   POLLING_STATION_DATA_ENTRY_SAVE_REQUEST_PATH,
 } from "@/types/generated/openapi";
 
-import { DataEntryState, DataEntryStateAndActions } from "../types/types";
+import { DataEntryStateAndActions } from "../types/types";
 import {
   onDeleteDataEntry,
   onFinaliseDataEntry,
@@ -24,13 +24,9 @@ export default function useDataEntry(
   election: Required<Election>,
   pollingStationId: number,
   entryNumber: number,
-  initialDataEntryState?: DataEntryState,
 ): DataEntryStateAndActions {
   const client = useApiClient();
-  const [state, dispatch] = useReducer(
-    dataEntryReducer,
-    initialDataEntryState || getInitialState(election, pollingStationId, entryNumber),
-  );
+  const [state, dispatch] = useReducer(dataEntryReducer, getInitialState(election, pollingStationId, entryNumber));
 
   // initial request to get the current data entry from the backend
   const saveRequestPath: POLLING_STATION_DATA_ENTRY_SAVE_REQUEST_PATH = `/api/polling_stations/${pollingStationId}/data_entries/${entryNumber}`;
